Reformat example plugin, clean it by a little.
